worker/pyproject.toml
[project]
name = "anywork-worker"
version = "0.1.0"
description = "AnyWork Agent Worker - nanobot-based containerized AI agent"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "nanobot-ai>=0.1.4",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.0.0",
    "pydantic>=2.0.0",
    "openai>=1.60.0",
    "anthropic>=0.42.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"
